requirements.txt
aws-cdk.core==1.134.0
aws-cdk.aws_ecr==1.134.0
aws-cdk.aws_codebuild==1.134.0
